# backend/openapi.yaml
openapi: 3.1.0
info:
  title: LangGraph Chatbot API
  description: AI-powered chatbot with conversation history using LangGraph and MongoDB
  version: 1.0.0
paths:
...
  /user_management/admin/me:
    get:
      tags:
      - Admin Profile
      summary: Read Admin Me
      description: "Get the current logged-in admin's profile details.\n\nReturns:\n\
        \    Admin profile information"
      operationId: read_admin_me_user_management_admin_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
      security:
      - OAuth2PasswordBearer: []
    put:
      tags:
      - Admin Profile
      summary: Update Admin Me
      description: "Update the current admin's profile.\n\nAllows updating name or\
        \ password. Passwords are automatically hashed.\nSends confirmation email\
        \ after successful update.\n\nArgs:\n    update_data: Fields to update (name,\
        \ password)\n    current_user: Authenticated admin from JWT token\n\nReturns:\n\
        \    Success message"
      operationId: update_admin_me_user_management_admin_me_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
    delete:
      tags:
      - Admin Profile
      summary: Delete Admin Me
      description: "Delete the authenticated admin's account permanently.\n\nRemoves\
        \ admin from database and sends confirmation email.\n\nArgs:\n    current_user:\
        \ Authenticated admin from JWT token\n\nReturns:\n    204 No Content on successful\
        \ deletion\n\nRaises:\n    HTTPException: If user not found"
      operationId: delete_admin_me_user_management_admin_me_delete
      responses:
        '204':
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
  /user_management/me:
    get:
      tags:
      - User Profile
      summary: Read User Me
      description: "Get the current logged-in user's profile details.\n\nReturns:\n\
        \    User profile information"
      operationId: read_user_me_user_management_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
      security:
      - OAuth2PasswordBearer: []
    put:
      tags:
      - User Profile
      summary: Update User Me
      description: "Update the current user's profile.\n\nAllows updating name or\
        \ password. Passwords are automatically hashed.\nSends confirmation email\
        \ after successful update.\n\nArgs:\n    update_data: Fields to update (name,\
        \ password)\n    current_user: Authenticated user from JWT token\n\nReturns:\n\
        \    Success message"
      operationId: update_user_me_user_management_me_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
    delete:
      tags:
      - User Profile
      summary: Delete User Me
      description: "Delete the authenticated user's account permanently.\n\nRemoves\
        \ user from database and sends confirmation email.\n\nArgs:\n    current_user:\
        \ Authenticated user from JWT token\n\nReturns:\n    204 No Content on successful\
        \ deletion\n\nRaises:\n    HTTPException: If user not found"
      operationId: delete_user_me_user_management_me_delete
      responses:
        '204':
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
  /user_management/admin/users:
    get:
      tags:
      - User Administration
      summary: List All Users
      description: "Get a list of all users (Admin only).\n\nReturns:\n    List of\
        \ all users without password hashes, including user_id"
      operationId: list_all_users_user_management_admin_users_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ReturnUser'
                type: array
                title: Response List All Users User Management Admin Users Get
      security:
      - OAuth2PasswordBearer: []
    post:
      tags:
      - User Administration
      summary: Create New User
      description: "Create a new user (Admin only).\n\nArgs:\n    user: User data\
        \ including email, password, name, and role\n\nReturns:\n    Created user\
        \ data including user_id\n\nRaises:\n    HTTPException: If email already registered"
      operationId: create_new_user_user_management_admin_users_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnUser'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
    delete:
      tags:
      - User Administration
      summary: Delete A User
      description: "Delete a user by email (Admin only).\n\nArgs:\n    user_to_delete:\
        \ Email of user to delete\n\nReturns:\n    204 No Content on successful deletion\n\
        \nRaises:\n    HTTPException: If user not found"
      operationId: delete_a_user_user_management_admin_users_delete
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteUser'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /user_management/admin/users/{email}:
    patch:
      tags:
      - User Administration
      summary: Update A User
      description: "Update a user's details by email (Admin only).\n\nAllows updating\
        \ name, role, is_active, is_bounced, is_complained.\nEmail cannot be changed.\n\
        \nArgs:\n    email: Email of the user to update\n    update_data: Fields to\
        \ update\n\nReturns:\n    Updated user data\n\nRaises:\n    HTTPException:\
        \ If user not found or no valid fields provided"
      operationId: update_a_user_user_management_admin_users__email__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: email
        in: path
        required: true
        schema:
          type: string
          title: Email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnUser'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /user_management/admin/all_messages:
    get:
      tags:
      - User Administration
      summary: Get All User Messages
      description: "Get a global list of all messages from all users (Admin only).\n\
        Sorted by latest message first.\n\nArgs:\n    skip: Number of records to skip\
        \ (pagination)\n    limit: Max number of records to return (default 50)\n\
        \    include_deleted: If True, include all messages (both deleted and active)\n\
        \    deleted_only: If True, return only deleted messages\n    \nReturns:\n\
        \    List of admin message logs"
      operationId: get_all_user_messages_user_management_admin_all_messages_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: include_deleted
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Include Deleted
      - name: deleted_only
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Deleted Only
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminMessageLog'
                title: Response Get All User Messages User Management Admin All Messages
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AdminMessageLog:
      properties:
        user_id:
          type: string
          title: User Id
        user_name:
          type: string
          title: User Name
        user_email:
          type: string
          title: User Email
        input_message:
          type: string
          title: Input Message
        ai_response:
          type: string
          title: Ai Response
        timestamp:
          type: string
          format: date-time
          title: Timestamp
        chat_id:
          type: string
          title: Chat Id
        session_id:
          type: string
          title: Session Id
        model_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Model Name
        deleted:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Deleted
          default: false
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
      type: object
      required:
      - user_id
      - user_name
      - user_email
      - input_message
      - ai_response
      - timestamp
      - chat_id
      - session_id
      title: AdminMessageLog
      description: 'Represents a flattened message log for the Admin Activity Feed.

        This is a denormalized record optimized for fast read/sort operations.'
    AdminUserUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        role:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
        is_bounced:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Bounced
        is_complained:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Complained
      type: object
      title: AdminUserUpdate
      description: Fields an admin can update on any user's account.
...
    CreateUser:
      properties:
        name:
          type: string
          title: Name
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          title: Password
        role:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
          default: user
      type: object
      required:
      - name
      - email
      - password
      title: CreateUser
      description: Payload for admin-created user accounts.
...
    DeleteUser:
      properties:
        email:
          type: string
          format: email
          title: Email
      type: object
      required:
      - email
      title: DeleteUser
      description: Payload for deleting a user by email.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MessageResponse:
      properties:
        message:
          type: string
          title: Message
      type: object
      required:
      - message
      title: MessageResponse
      description: Generic single-message response body for informational endpoints.
...
    ReturnUser:
      properties:
        user_id:
          type: string
          title: User Id
        name:
          type: string
          title: Name
        email:
          type: string
          format: email
          title: Email
        role:
          type: string
          title: Role
        auth_method:
          anyOf:
          - type: string
          - type: 'null'
          title: Auth Method
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
          default: true
        is_bounced:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Bounced
          default: false
        is_complained:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Complained
          default: false
        chat_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Chat Count
          default: 0
        total_tokens:
          anyOf:
          - type: integer
          - type: 'null'
          title: Total Tokens
          default: 0
        models_used:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Models Used
          default: []
      type: object
      required:
      - user_id
      - name
      - email
      - role
      title: ReturnUser
      description: Full user record returned in admin user management endpoints.
...
    UserResponse:
      properties:
        name:
          type: string
          title: Name
        email:
          type: string
          format: email
          title: Email
        role:
          type: string
          title: Role
        created_at:
          type: string
          format: date-time
          title: Created At
        is_active:
          type: boolean
          title: Is Active
      type: object
      required:
      - name
      - email
      - role
      - created_at
      - is_active
      title: UserResponse
      description: Public user profile returned from GET /me endpoints.
    UserUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        password:
          anyOf:
          - type: string
          - type: 'null'
          title: Password
      type: object
      title: UserUpdate
      description: Fields available for a user to update on their own profile.
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: auth/login/user